requirements.txt
tensorflow-gpu==1.12.0
sklearn
keras
imutils
opencv-python
matplotlib
